refactor(persistence): delete deprecation stub; extract serializers + time_info (2.8.3)
- mud/persistence.py removed; surfaces split:
  - mud/db/serializers.py: _serialize_*/_deserialize_* helpers + ObjectSave/PlayerSave
    used by mud/account/account_manager.py and mud/models/character.py:from_orm
  - mud/world/time_persistence.py: time_info save/load
- Updated importers (account_manager, character.from_orm, tests/test_time_persistence)
- CROSS_FILE_INVARIANTS_TRACKER INV-008 row + SESSION_STATUS reflect the split
- pyproject 2.8.2 -> 2.8.3 (patch; no API change visible to users)

INV-008 enforcement test + DB-canonical round-trip + character creation runtime + time_persistence: 26/26 green.
