+++ mud/models/character.py
@@ -1168,10 +1168,10 @@ def from_orm(db_char: DBCharacter) -> Character:
     # colours
     saved_colours = getattr(db_char, "colours", None)
     if saved_colours and isinstance(saved_colours, dict):
-        from mud.persistence import _apply_colour_table, _normalize_int_list
+        from mud.db.serializers import _apply_colour_table, _normalize_int_list
         _apply_colour_table(pcdata, saved_colours)
     else:
-        from mud.persistence import _normalize_int_list
+        from mud.db.serializers import _normalize_int_list
 
     # mod_stat and armor
     saved_mod_stat = getattr(db_char, "mod_stat", None)
@@ -1189,7 +1189,7 @@ def from_orm(db_char: DBCharacter) -> Character:
     # in the DB-canonical path — it only restores prototype defaults.
     inventory_state = getattr(db_char, "inventory_state", None)
     if inventory_state and isinstance(inventory_state, list):
-        from mud.persistence import _deserialize_object, ObjectSave
+        from mud.db.serializers import _deserialize_object, ObjectSave
         from mud.models.json_io import dataclass_from_dict
         restored_inventory = []
         for obj_dict in inventory_state:
@@ -1204,7 +1204,7 @@ def from_orm(db_char: DBCharacter) -> Character:
 
     equipment_state = getattr(db_char, "equipment_state", None)
     if equipment_state and isinstance(equipment_state, dict):
-        from mud.persistence import _deserialize_object, ObjectSave
+        from mud.db.serializers import _deserialize_object, ObjectSave
         from mud.models.json_io import dataclass_from_dict
         restored_equipment: dict[str, Object] = {}
         for slot, obj_dict in equipment_state.items():
@@ -1220,7 +1220,7 @@ def from_orm(db_char: DBCharacter) -> Character:
     # --- INV-008 Phase 2: deserialize pet JSON blob ---
     pet_state = getattr(db_char, "pet_state", None)
     if pet_state and isinstance(pet_state, dict):
-        from mud.persistence import _deserialize_pet
+        from mud.db.serializers import _deserialize_pet
         try:
             pet = _deserialize_pet(pet_state, char)
             if pet is not None:
