+++ mud/world/time_persistence.py
@@ -1,0 +1,61 @@
+from __future__ import annotations
+
+"""Time-info persistence — save/load the global clock to/from disk.
+
+Extracted from mud.persistence (deprecated stub) in 2.8.3.
+ROM C reference: src/db.c — time_info is initialised at startup and
+ticked each game loop; we persist it across restarts via a JSON file.
+"""
+
+import os
+from dataclasses import dataclass
+from pathlib import Path
+
+from mud.models.json_io import dump_dataclass, load_dataclass
+from mud.time import Sunlight, time_info
+
+TIME_FILE = Path("data/time.json")
+
+
+@dataclass
+class TimeSave:
+    hour: int
+    day: int
+    month: int
+    year: int
+    sunlight: int
+
+
+def save_time_info() -> None:
+    """Persist global time_info to TIME_FILE (atomic write)."""
+    TIME_FILE.parent.mkdir(parents=True, exist_ok=True)
+    data = TimeSave(
+        hour=time_info.hour,
+        day=time_info.day,
+        month=time_info.month,
+        year=time_info.year,
+        sunlight=int(time_info.sunlight),
+    )
+    tmp_path = TIME_FILE.with_suffix(".tmp")
+    with tmp_path.open("w") as f:
+        dump_dataclass(data, f, indent=2)
+        f.flush()
+        os.fsync(f.fileno())
+    os.replace(tmp_path, TIME_FILE)
+
+
+def load_time_info() -> None:
+    """Load global time_info from TIME_FILE if present."""
+    if not TIME_FILE.exists():
+        return
+    with TIME_FILE.open() as f:
+        data = load_dataclass(TimeSave, f)
+    time_info.hour = data.hour
+    time_info.day = data.day
+    time_info.month = data.month
+    time_info.year = data.year
+    try:
+        time_info.sunlight = Sunlight(data.sunlight)
+    except Exception:
+        # Fallback if invalid value
+        time_info.sunlight = Sunlight.DARK
